+++ Userland/SampleCodeModule/pipes.c
@@ -6,11 +6,11 @@ int newPipe(){
 }
 
 void sendMessagePipe(int pipe, char* msg, int length){
-  systemCall(23, pipe, msg, length,0,0);
+  systemCall(23, pipe, (uint64_t)msg, length,0,0);
 }
 
 void receiveMessagePipe(int pipe, char* dest, int length){
-  systemCall(24, pipe, dest, length,0,0);
+  systemCall(24, pipe, (uint64_t)dest, length,0,0);
 }
 
 void deletePipe(){
